requirements.txt
flask
openai
langchain
PyYAML
faiss-cpu
twilio
colorama
redis
rich
sqlalchemy[mypy]
loguru